Change directory structure

+++ script/install/install_package.sh
@@ -3,17 +3,19 @@
 
 # Show Help Info
 if [[ $1 = "-h" || $1 = "--help" ]];then
+    echo -e "$0 [LIST]"
     echo -e "Install packages"
     exit 0
 fi
 
 # Termux
-if [ -n "$TERMUX" ];then
+if [[ -n "$TERMUX" && -z "$POINTLESS" ]];then
     install_pkg wget bash
     wget https://its-pointless.github.io/setup-pointless-repo.sh -O setup-pointless-repo.sh
     bash setup-pointless-repo.sh
     rm -f setup-pointless-repo.sh pointless.gpg
 fi
 
-install_pkg $(pkg_list)
+PKGS=$(pkg_list $1)
+[ -n "$PKGS" ] && install_pkg $PKGS
